# specs/openapi.yaml
openapi: 3.1.0
info:
  title: AIOps Console API
  version: 0.1.0-alpha
  description: |
    AI 原生后台管理系统 API，覆盖 Prompt 管理、Agent 编排、知识库、模型路由、
    对话分析与评估套件六大领域。

    遵循 Agentic Engineering 哲学：Spec 即契约，类型即文档。
  contact:
    name: AIOps Console Team
    url: https://github.com/aiops/console
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.aiops.local/api/v1
    description: Production
...
paths:
  /prompts:
    get:
      tags: [Prompts]
      summary: 列出 Prompt
      operationId: listPrompts
      parameters:
        - name: q
          in: query
          description: 名称或描述模糊搜索关键词
          schema:
            type: string
        - name: limit
          in: query
          description: 返回条数上限
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: 分页偏移量
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Prompt 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromptOut'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Prompts]
      summary: 创建 Prompt
      operationId: createPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /prompts/{prompt_id}:
    parameters:
      - $ref: '#/components/parameters/PromptId'
    get:
      tags: [Prompts]
      summary: 获取 Prompt 详情
      operationId: getPrompt
      responses:
        '200':
          description: Prompt 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptOut'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Prompts]
      summary: 更新 Prompt 元数据
      operationId: updatePrompt
      description: 仅更新 name 与 description，不产生新版本。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 128
                description:
                  type: string
                  maxLength: 512
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptOut'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Prompts]
      summary: 删除 Prompt
      operationId: deletePrompt
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /prompts/{prompt_id}/versions:
    parameters:
      - $ref: '#/components/parameters/PromptId'
    get:
      tags: [Prompts]
      summary: 列出 Prompt 的所有版本
      operationId: listPromptVersions
      responses:
        '200':
          description: 版本列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromptVersionOut'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Prompts]
      summary: 创建 Prompt 新版本
      operationId: createPromptVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 65536
                variables:
                  type: array
                  items:
                    type: string
                    pattern: '^[a-zA-Z0-9_]+$'
                  default: []
      responses:
        '201':
          description: 新版本创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersionOut'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /prompts/{prompt_id}/versions/{version_id}/rollback:
    parameters:
      - $ref: '#/components/parameters/PromptId'
      - $ref: '#/components/parameters/VersionId'
    post:
      tags: [Prompts]
      summary: 回滚到指定版本
      operationId: rollbackPromptVersion
      description: 将指定版本设为当前版本，并产生新的版本号。
      responses:
        '200':
          description: 回滚成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptOut'
        '404':
          $ref: '#/components/responses/NotFound'
  /prompts/{prompt_id}/diff:
    parameters:
      - $ref: '#/components/parameters/PromptId'
    get:
      tags: [Prompts]
      summary: 对比两个版本的差异
      operationId: diffPromptVersions
      parameters:
        - name: from_version
          in: query
          required: true
          schema:
            type: integer
        - name: to_version
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 版本差异
          content:
            application/json:
              schema:
                type: object
                properties:
                  from_version:
                    type: integer
                  to_version:
                    type: integer
                  diff:
                    type: string
                    description: unified diff 文本
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  parameters:
    PromptId:
      name: prompt_id
      in: path
      required: true
      schema:
        type: integer
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
    PromptCreate:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: Prompt 名称
        description:
          type: string
          maxLength: 512
          description: Prompt 描述
        content:
          type: string
          minLength: 1
          maxLength: 65536
          description: Prompt 模板内容（最大 64KB）
        variables:
          type: array
          items:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
          description: 模板变量名列表
          default: []
    PromptVersionOut:
      type: object
      required: [id, version_num, content, variables, created_at]
      properties:
        id:
          type: integer
        version_num:
          type: integer
          description: 版本号，从 1 递增
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    PromptOut:
      type: object
      required: [id, name, current_version, version_count, created_at, updated_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        current_version:
          $ref: '#/components/schemas/PromptVersionOut'
        version_count:
          type: integer
          description: 版本总数（最大 100）
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...